Advancements, path unlocking +Added advancements. There are now 88 of them, more are planned. *Changed how path unlocking works. Default behavior (configurable) is that after you unlock your first path, you need to complete some advancement related to that path to unlock another one. Path unlocking is now a button in the Detailed Crafter UI instead of a weird pseudo-recipe thing.

+++ src/main/java/com/Da_Technomancer/crossroads/render/BeamToolOverlay.java
@@ -87,7 +87,7 @@ public class BeamToolOverlay implements LayeredDraw.Layer{
 				builder.addVertex(matrix.last().pose(), 117, 0, -3).setUv(1, 0);
 				builder.addVertex(matrix.last().pose(), 0, 0, -3).setUv(0, 0);
 
-				byte[] settings = BeamUsingItem.getSetting(mainStack);
+				int[] settings = BeamUsingItem.getSetting(mainStack);
 				for(int i = 0; i < 4; i++){
 					float fullness = (float) settings[i] / 8;
 					int extension = (int) (72 * fullness);
